refactor(teleport): satisfy the layer-stack gate main introduced
The rebase brought `check-layer-back-edges`, which replaced the ui-only gate
with the full stack (fs → shell/git → cdp → tools → core → scoops → ui). This
branch tripped it three ways.

**`tab-teleport.ts` moved to `scoops/tray-leader/`.** In `shell/` it imported
`cdp/` and `core/logger.js` — two up-the-stack edges. Its only consumers are
`ui/wc/wc-browser.ts` and `scoops/tray-leader/tab-teleport-router.ts`, and the
tray is where it belongs anyway: beside the router that drives it. From there
every import points down, including `teleport-storage.js` back in `shell/`. The
grandfathered Okta watcher next door has the same two edges, so growing the
baseline was available — the gate says move the code instead, and here the move
is also the better home.

**`host-command.ts` is no longer touched.** The change was six lines: three
optional fields on `ConnectedFollowerInfo`. That file carries 7 grandfathered
back-edges, and the boy-scout gate requires a PR touching such a file to pay ALL
of them down — a refactor of the host command, unrelated to teleport. The fields
now live on `TeleportFollowerInfo` in `teleport-follower-shim.ts`, the shell-layer
module that consumes them; `wc-tray.ts` produces that type (ui → shell points
down) and the extra keys ride through the shim regardless, since it is JSON.

**Baseline ratcheted, downward.** Three entries were stale because this branch
had already paid them off: moving the permissions-surface registry to `core/`
cleaned `providers/oauth-service.ts`, `speech/composer-speech.ts` and
`speech/hear.ts`. 151 → 148 back-edges across 88 files.

Verification: lint (all 7), typecheck, 13,929 tests, coverage floors (each
package gate individually), both builds, bundle size, and the debt gate now
reporting 0 touched debt files.
